requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
bcrypt==5.0.0
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
h11==0.16.0
httptools==0.6.4
idna==3.10
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing==3.7.4.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
